add "conn" and "err" states to display
currently it's easy to get confused between the various states the device can be in wrt wifi disabled/connecting/AP-setup,
which all display "----", so this adds "conn" to the display instead during the AP-setup mode. in addition, it's useful
to be able to tell when the device can't read from the sensor, so a 2s timeout is added along with "err" on the display
in case no valid data was read.